CORE-10411: Deleteing the media only when there is no usage of file.

+++ docroot/modules/commerce/acq_commerce/modules/acq_sku/src/Entity/SKU.php
@@ -369,6 +369,23 @@ class SKU extends ContentEntityBase implements SKUInterface {
     $file_name_array[] = $extension;
     $file_name = implode('.', $file_name_array);
 
+    /** @var \Drupal\file\FileUsage\FileUsageInterface $file_usage */
+    $file_usage = \Drupal::service('file.usage');
+
+    // Check if file already exists in the directory.
+    if (file_exists($directory . '/' . $file_name)) {
+      // If file exists in directory, check if file entity exists.
+      $file_storage = \Drupal::entityTypeManager()->getStorage('file');
+      $files = reset($file_storage->loadByProperties(['uri' => $directory . '/' . $file_name]));
+      if (!empty($files) && $files instanceof FileInterface) {
+        // Release the lock now.
+        $lock->release($lock_key);
+        // Add file usage.
+        $file_usage->add($files, $this->getEntityTypeId(), $this->getEntityTypeId(), $this->id());
+        return $files;
+      }
+    }
+
     // Save the file as file entity.
     /** @var \Drupal\file\Entity\File $file */
     if ($file = file_save_data($file_data, $directory . '/' . $file_name, FILE_EXISTS_REPLACE)) {
@@ -380,8 +397,6 @@ class SKU extends ContentEntityBase implements SKUInterface {
         $lock->release($lock_key);
       }
 
-      /** @var \Drupal\file\FileUsage\FileUsageInterface $file_usage */
-      $file_usage = \Drupal::service('file.usage');
       $file_usage->add($file, $this->getEntityTypeId(), $this->getEntityTypeId(), $this->id());
 
       return $file;
@@ -799,7 +814,14 @@ class SKU extends ContentEntityBase implements SKUInterface {
       /** @var \Drupal\acq_sku\Entity\SKU $entity */
       foreach ($entity->getMedia(FALSE) as $media) {
         if ($media['file'] instanceof FileInterface) {
-          $media['file']->delete();
+          /** @var \Drupal\file\FileUsage\FileUsageInterface $file_usage */
+          $file_usage = \Drupal::service('file.usage');
+          // Delete usage of file.
+          $file_usage->delete($media['file'], $entity->getEntityTypeId(), $entity->getEntityTypeId(), $entity->id());
+          // If no usage of the image, delete it.
+          if (empty($file_usage->listUsage())) {
+            $media['file']->delete();
+          }
         }
       }
     }
